Remove gammapy/cube/utils.py
